Update custom-session-next-auth example
