+++ examples/custom-session-next-auth/session.ts
@@ -1,7 +1,6 @@
-import { getContext } from '@keystone-6/core/context'
 import { getServerSession } from 'next-auth/next'
 import type { DefaultJWT } from 'next-auth/jwt'
-import type { DefaultSession, DefaultUser } from 'next-auth'
+import type { DefaultSession } from 'next-auth'
 import GithubProvider from 'next-auth/providers/github'
 import type { Context } from '.keystone/types'
 
@@ -12,51 +11,10 @@ import type { Context } from '.keystone/types'
 // WARNING: you need to change this
 const sessionSecret = '-- DEV COOKIE SECRET; CHANGE ME --'
 
-let _keystoneContext: Context = (globalThis as any)._keystoneContext
-
-async function getKeystoneContext() {
-  if (_keystoneContext) return _keystoneContext
-
-  // TODO: this could probably be better
-  _keystoneContext = getContext(
-    (await import('./keystone')).default,
-
-    // WARNING: this is only needed for our monorepo examples, dont do this
-    await import('myprisma')
-    // await import('@prisma/client') // <-- do this
-  )
-  if (process.env.NODE_ENV !== 'production') {
-    ;(globalThis as any)._keystoneContext = _keystoneContext
-  }
-  return _keystoneContext
-}
-
 // see https://next-auth.js.org/configuration/options for more
 export const nextAuthOptions = {
   secret: sessionSecret,
   callbacks: {
-    async signIn({ user }: { user: DefaultUser }) {
-      // console.error('next-auth signIn', { user, account, profile });
-      const sudoContext = (await getKeystoneContext()).sudo()
-
-      // check if the user exists in keystone
-      const author = await sudoContext.query.Author.findOne({
-        where: { authId: user.id },
-      })
-
-      // if not, sign up
-      if (!author) {
-        await sudoContext.query.Author.createOne({
-          data: {
-            authId: user.id,
-            name: user.name,
-          },
-        })
-      }
-
-      return true // accept the signin
-    },
-
     async session({
       session,
       token,
